Extract HTML code to create popup sceens to its own class. Corrected some types to number
